feature(exchange): adding encryptus
